[TC] Regenerated demo for tc_tc_capture_mode_sam_a5d2_xult_iar
MH3-67223

+++ apps/tc/tc_capture_mode/firmware/src/config/sam_a5d2_xult_iar/peripheral/tc/plib_tc0.c
@@ -101,13 +101,13 @@ uint32_t TC0_CH2_CaptureFrequencyGet( void )
 /* Read last captured value of Capture A */
 uint32_t TC0_CH2_CaptureAGet (void)
 {
-    return TC0_REGS->TC_CHANNEL[2].TC_RA;
+    return (uint32_t)TC0_REGS->TC_CHANNEL[2].TC_RA;
 }
 
 /* Read last captured value of Capture B */
 uint32_t TC0_CH2_CaptureBGet (void)
 {
-    return TC0_REGS->TC_CHANNEL[2].TC_RB;
+    return (uint32_t)TC0_REGS->TC_CHANNEL[2].TC_RB;
 }
 
 TC_CAPTURE_STATUS TC0_CH2_CaptureStatusGet(void)
